Refactor app.py to integrate GUI functionality and enhance security monitoring features. Removed gui_app.py and consolidated code into app.py, which now includes real-time camera monitoring, threat analysis, and Discord notification capabilities. Updated README.md to reflect changes in usage instructions and application structure.
